basit

+++ app/Http/Controllers/WEB/Admin/Slider1Controller.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\WEB\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Slider1;
-use App\Models\Product;
+use App\Models\Brand;
 use Image;
 use File;
 use Stichoza\GoogleTranslate\GoogleTranslate;
@@ -23,7 +23,7 @@ class Slider1Controller extends Controller
     }
 
     public function create(){
-        $products = Product::where(['status' => 1])->select('id','name','slug')->get();
+        $products = Brand::where(['status' => 1])->select('id','name','slug')->get();
         return view('admin.slider1.create_slider', compact('products'));
     }
 
@@ -35,9 +35,9 @@ class Slider1Controller extends Controller
             'product_slug' => 'required',
             'status' => 'required',
             'serial' => 'required',
-            'title_one' => 'required',
-            'title_two' => 'required',
-            'badge' => 'required',
+            // 'title_one' => 'required',
+            // 'title_two' => 'required',
+            // 'badge' => 'required',
         ];
         $customMessages = [
             'slider_image.required' => trans('admin_validation.Slider image is required'),
@@ -60,10 +60,17 @@ class Slider1Controller extends Controller
                 ->save(public_path().'/'.$slider_image);
             $slider->image = $slider_image;
         }
-
-        $slider->badge_ar = $tr->translate($request->badge);
-        $slider->title_one_ar = $tr->translate($request->title_one);
-        $slider->title_two_ar = $tr->translate($request->title_two);
+        if($request->slider_image_ar){
+            $extention = $request->slider_image_ar->getClientOriginalExtension();
+            $slider_image = 'slider'.date('-Y-m-d-h-i-s-').rand(999,9999).'.'.$extention;
+            $slider_image = 'uploads/custom-images/'.$slider_image;
+            Image::make($request->slider_image_ar)
+                ->save(public_path().'/'.$slider_image);
+            $slider->image_ar = $slider_image;
+        }
+        $slider->badge_ar = $request->badge ? $tr->translate($request->badge):'';
+        $slider->title_one_ar = $request->title_one ? $tr->translate($request->title_one) : '';
+        $slider->title_two_ar = $request->title_two ? $tr->translate($request->title_two) : '';
 
         $slider->product_slug = $request->product_slug;
         $slider->serial = $request->serial;
@@ -85,7 +92,7 @@ class Slider1Controller extends Controller
 
     public function edit($id){
         $slider = Slider1::find($id);
-        $products = Product::where(['status' => 1])->select('id','name','slug')->get();
+        $products = Brand::where(['status' => 1])->select('id','name','slug')->get();
         return view('admin.slider1.edit_slider', compact('slider','products'));
     }
 
@@ -94,9 +101,9 @@ class Slider1Controller extends Controller
             'product_slug' => 'required',
             'status' => 'required',
             'serial' => 'required',
-            'title_one' => 'required',
-            'title_two' => 'required',
-            'badge' => 'required',
+            // 'title_one' => 'required',
+            // 'title_two' => 'required',
+            // 'badge' => 'required',
 
         ];
         $customMessages = [
@@ -115,6 +122,19 @@ class Slider1Controller extends Controller
             Image::make($request->slider_image)
                 ->save(public_path().'/'.$slider_image);
             $slider->image = $slider_image;
+            $slider->save();
+            if($existing_slider){
+                if(File::exists(public_path().'/'.$existing_slider))unlink(public_path().'/'.$existing_slider);
+            }
+        }
+        if($request->slider_image_ar){
+            $existing_slider = $slider->image_ar;
+            $extention = $request->slider_image_ar->getClientOriginalExtension();
+            $slider_image = 'slider'.date('-Y-m-d-h-i-s-').rand(999,9999).'.'.$extention;
+            $slider_image = 'uploads/custom-images/'.$slider_image;
+            Image::make($request->slider_image_ar)
+                ->save(public_path().'/'.$slider_image);
+            $slider->image_ar = $slider_image;
             $slider->save();
             if($existing_slider){
                 if(File::exists(public_path().'/'.$existing_slider))unlink(public_path().'/'.$existing_slider);
